push_swap added

+++ ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: smeza-ro <smeza-ro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/02 17:39:39 by smeza-ro          #+#    #+#             */
-/*   Updated: 2026/02/16 14:44:03 by smeza-ro         ###   ########.fr       */
+/*   Updated: 2026/03/13 10:10:59 by smeza-ro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-int	ft_atoi(const char *nptr)
+long int	ft_atoi(const char *nptr)
 {
-	int	i;
-	int	sign;
-	int	res;
+	int			i;
+	int			sign;
+	long int	res;
 
 	i = 0;
 	sign = 1;
@@ -34,6 +34,5 @@ int	ft_atoi(const char *nptr)
 		res = res * 10 + (nptr[i] - 48);
 		i++;
 	}
-	//if (!(nptr[i] >= '0' && nptr[i] <= '9'))
 	return (sign * res);
 }
